refactor: move all xxx-director into directors dir
