+++ internal/apps/ai-proxy/dependent_filters.go
@@ -18,7 +18,6 @@ import (
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/assets"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/audit-after-llm-director"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/audit-before-llm-director"
-	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/azure-director"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/body-size-limit"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/context"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/context-assistant"
@@ -29,13 +28,14 @@ import (
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/context-image"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/context-responses"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/dashscope-director"
+	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/directors/azure-director"
+	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/directors/openai-compatible-director"
+	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/directors/openai-director"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/erda-auth"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/extra-body"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/finalize"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/initialize"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/log-http"
-	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/openai-compatible-director"
-	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/openai-director"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/openai-v1-models"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/prometheus-collector"
 	_ "github.com/erda-project/erda/internal/apps/ai-proxy/filters/rate-limit"
